Console Render Bug caused by touch drag event
With starting the console the touch drag event is active without even touching the display.
This causes to render the console at any time it is displayed, which consumes alot of battery and heats up the device.
What the patch does is resetting the touch drag state to ensure it is only re-rendering the console when it it really needed.